+++ src/main/java/roomescape/domain/theme/ThemeRepository.java
@@ -1,9 +1,8 @@
-package roomescape.domain.theme.repository;
+package roomescape.domain.theme;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
-import roomescape.domain.theme.Theme;
 
 public interface ThemeRepository {
     Theme save(Theme theme);
