+++ src/main/java/roomescape/infrastructure/JdbcMemberRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain.member.repository;
+package roomescape.infrastructure;
 
 import java.util.Optional;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -6,6 +6,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.member.Member;
+import roomescape.domain.member.MemberRepository;
 import roomescape.domain.member.Role;
 
 @Repository
