+++ src/test/java/roomescape/infrastructure/JdbcMemberRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.member.repository;
+package roomescape.infrastructure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.domain.member.Member;
+import roomescape.domain.member.MemberRepository;
 
 @JdbcTest
 class JdbcMemberRepositoryTest {
