+++ src/test/java/roomescape/infrastructure/JdbcReservationTimeRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.time.repository;
+package roomescape.infrastructure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -12,6 +12,7 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.domain.time.ReservationTime;
+import roomescape.domain.time.ReservationTimeRepository;
 import roomescape.fixture.ReservationTimeFixture;
 
 @JdbcTest
